Update source and test driver for running test262 on windows
* Update test262 driver for python3 and windows
* Update DateFormat for windows
* Fix compiler warnings on Win64

+++ src/runtime/ObjectStructure.h
@@ -264,7 +264,7 @@ private:
             return 1;
         }
         size_t base = FAST_LOG2_UINT(newSize);
-        return 1 << (base + 1);
+        return size_t(1) << (base + 1);
     }
 
     ObjectStructureItemTightVector m_properties;
